Uploaded function to calculate bounds of elastic constants
This function calculates the bounds on the elastic constants  of
statistically isotropic and homogeneous polycrystals containing a single
phase of spherical crystallites. This procedure is based on the theory
described in the article, "Bounds and self-consistent estimates of the
elastic constants of polycrystals" by Christopher M. Kube and Andrea P.
Arguelles.